Initial
PHP version, including working example. Translation mode (mode 3) not
html smart: it replaces anything it recognizes that is surrounded by
whitespace (conversely, it will not replace something bordered by an
angle bracket). I'll fix that later. Also, whitespace type is not
preserved: tabs and line breaks will become spaces. Currently unsure if
I want to change that in the future, as it would overcomplicate it.